Add bot command handling
Allows messages prefixed with known bot command characters to be sent
to both Discord and IRC without a nickname prefix, as most bots
expect their command to be the first word in a message. For
identification, the user who typed the command is sent in a separate
message before sending the non-prefixed message.